Calcul du score

+++ geopictures_back/sql/schema.sql
@@ -259,7 +259,7 @@ create table if not exists photo_joueur
     joueur_id bigint NOT NULL,
     photo_id bigint not null,
     image varchar(255) not null,
-    score int not null,
+    score numeric(19, 2) not null,
     succes_gps boolean not null,
     succes_globale boolean not null,
     constraint fk_photo_joueur_joueur
